Renamed types folder to utils, changed command register script to typescript
