+++ src/commands/poll/delete-poll.ts
@@ -1,6 +1,6 @@
 import { CommandInteraction, SlashCommandBuilder } from 'discord.js';
 import { deletePoll, verifyDeletePermission } from '../../database/poll-db';
-import { ReplyStatus } from '../../types/discordTypes';
+import { ReplyStatus } from '../../utils/discordUtils';
 
 export const data = new SlashCommandBuilder()
   .setName('delete-poll')
